backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-docx==1.1.0
PyMuPDF==1.23.8
pandas==2.1.3
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.12.1
redis==5.0.1
jieba==0.42.1
python-dateutil==2.8.2
aiofiles==23.2.1
python-dotenv==1.0.0
